Allowing vitess Online DDL in Vitess to work with foreign keys.
We wish to solve the issue of an Online DDL cut-over. We distinguish
two scenarios: migrating a foreign key (FK) _parent_ table, and
migrating a FK _child_ table.

Since we rename the parent table away, we must iterate its children.
Standard MySQL behavior is for children to follow the renamed parent
into its new name. In fact, internally to InnoDB, the foreign key
relationships just maintain their existing pointer referenced between
the related tables. All that MySQL does is to modify the children's
`CREATE TABLE` statement to point to the new name.

**What we do**: we avoid changing the children's `CREATE TABLE`
statement, and we ensure to refresh (reset) the FK relationship, such
that the child points to the new parent (new parent assumed the old
parent's name).

This seems to work well under stress tests.

As a reminder: if we are migrating a child table, then we create a
shadow table, that will eventually replace the child. If that shadow
table references the parent, then in the duration of the migration, and
while the shadow table is not in full sync with the original child
table, the shadow table may actually cause valid queries running on the
parent table to fail. Specifically, a `DELETE FROM parent_table
WHERE...` can fail if the shadow table has rows matching the rows
deleted on the parent.

In this PR we let the shadow table reference the same parent. We tell
InnoDB to ignore FK relationships on tables which match internal Vitess
names, namely names like:

- `_84371a37_6153_11eb_9917_f875a4d24e90_20210128122816_vrepl` (Online
DDL vreplication shadow table format)
- `_vt_HOLD_6ace8bcef73211ea87e9f875a4d24e90_20200915120410` (GC name
format)
- `_vt_DROP_6ace8bcef73211ea87e9f875a4d24e90_20200915120410` (GC name
format)

With the new rules, InnoDB treats any FK relationship involving
Vitess-internal tables as if we set `FOREIGN_KEY_CHECKS=0`, specific to
those tables.

Examples:

- If `parent_table` is parent, and
`_84371a37_6153_11eb_9917_f875a4d24e90_20210128122816_vrepl` is child,
then
  - It is possible to `INSERT` any rows to
`_84371a37_6153_11eb_9917_f875a4d24e90_20210128122816_vrepl`, whether
they do or do not match rows in `parent_table`
  - It is possible to `DELETE` or `UPDATE` rows on `parent_table`,
irrespective of whether matching rows do or do not exist in
`_84371a37_6153_11eb_9917_f875a4d24e90_20210128122816_vrepl`. There
will not even be a check.
  - `CASCADE` and `SET NULL` are never attempted on
`_84371a37_6153_11eb_9917_f875a4d24e90_20210128122816_vrepl`
- If `_84371a37_6153_11eb_9917_f875a4d24e90_20210128122816_vrepl` is
the parent, and `child_table` is child, then
  - It is possible to `DELETE` or `UPDATE` any rows on
`_84371a37_6153_11eb_9917_f875a4d24e90_20210128122816_vrepl`,
regardless of whether matching rows do or do not exist in `child_table`
  - It is possible to `DROP TABLE
_84371a37_6153_11eb_9917_f875a4d24e90_20210128122816_vrepl`

A use case might be:
```sql
SET rename_table_preserve_foreign_key=1;
rename table a to b, b to c, c to a;
SET rename_table_preserve_foreign_key=0;
```

This makes it simple to verify whether a MySQL server supports this
feature, as in:
```sql
> show variables like 'rename_table_preserve_foreign_key';
+-----------------------------------+-------+
| Variable_name                     | Value |
+-----------------------------------+-------+
| rename_table_preserve_foreign_key | OFF   |
+-----------------------------------+-------+
```

A server which does not support this variable will just return an empty
result set, and without any error.

+++ sql/sql_table.cc
@@ -2455,7 +2455,18 @@ static bool rm_table_check_fks(THD *thd, Drop_tables_ctx *drop_ctx) {
       return true;
     }
 
+    Table_name_inspector inspector(table->table_name);
+    if (inspector.skip_fk_checks()) {
+      continue;
+    }
+
     for (const dd::Foreign_key_parent *fk : table_def->foreign_key_parents()) {
+      // Whether we drop a table or a database, it's fine to ignore vitess-internal children tables:
+      Table_name_inspector child_inspector(fk->child_table_name().c_str());
+      if (child_inspector.skip_fk_checks()) {
+        // The child table can be ignored.
+        continue;
+      }
       if (drop_ctx->drop_database) {
         /*
           In case of DROP DATABASE list of tables to be dropped can be huge.
@@ -9827,6 +9838,58 @@ static bool check_fk_children_after_parent_def_change(
   @param parent_table_db    Old schema name.
   @param parent_table_name  Old table name.
   @param hton               Handlerton for table's storage engine.
+
+  @retval operation outcome, false if no error.
+*/
+static bool reload_fk_children_after_parent_rename_preserve(
+    THD *thd,
+    const char *parent_table_db,
+    const char *parent_table_name,
+    handlerton *hton) {
+  Normalized_fk_children fk_children;
+  if (fetch_fk_children_uncached_uncommitted_normalized(
+          thd, parent_table_db, parent_table_name,
+          ha_resolve_storage_engine_name(hton), &fk_children))
+    return true;
+
+  for (auto fk_children_it : fk_children) {
+    const char *schema_name = fk_children_it.first.c_str();
+    const char *table_name = fk_children_it.second.c_str();
+
+    if (my_strcasecmp(table_alias_charset, schema_name, parent_table_db) == 0 &&
+        my_strcasecmp(table_alias_charset, table_name, parent_table_name) ==
+            0) {
+      // self reference
+      continue;
+    }
+
+    dd::Table *child_table_def = nullptr;
+
+    if (thd->dd_client()->acquire_for_modification(schema_name, table_name,
+                                                   &child_table_def))
+      return true;
+
+    assert(child_table_def != nullptr);
+
+    mysql_ha_flush_table(thd, schema_name, table_name);
+    close_all_tables_for_name(thd, schema_name, table_name, false);
+
+    if (hton->dict_cache_reset) {
+      hton->dict_cache_reset(schema_name, table_name);
+    }
+  }
+
+  return false;
+}
+
+/**
+  Update the referenced schema- and/or table name for the referencing tables
+  when the referenced table is renamed.
+
+  @param thd                Thread handle.
+  @param parent_table_db    Old schema name.
+  @param parent_table_name  Old table name.
+  @param hton               Handlerton for table's storage engine.
   @param new_db             New schema name.
   @param new_table_name     New table name.
 
@@ -15720,6 +15783,24 @@ bool adjust_fks_for_rename_table(THD *thd, const char *db,
   if (adjust_fk_parents(thd, new_db, new_table_name, true, nullptr))
     return true;
 
+  return false;
+}
+
+
+bool adjust_fks_for_rename_table_with_preserve_fk(THD *thd, const char *db,
+                                 const char *table_name, const char *new_db,
+                                 const char *new_table_name,
+                                 handlerton *hton) {
+  dd::Table *new_table = nullptr;
+
+  if (thd->dd_client()->acquire_for_modification(new_db, new_table_name, &new_table))
+    return true;
+
+  assert(new_table != nullptr);
+
+  if (reload_fk_children_after_parent_rename_preserve(thd, db, table_name, hton)) {
+    return true;
+  }
   return false;
 }
